alsa_conformance_test: Consider delay when calculating the played frames
Use the delay value to instead of frames_left to calculate the number of
frames played out.
If the driver does not report delay (delay = 0) then the value of the delay
is exactly what the frames_left is since all frames considered to be played
out right away.
The delay means that despite that the DMA has moved the data, it is not yet
played out.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>